项目结构修改 CurveNavigationView - 代码修改 SimpleCRC - 开发
