+++ core/src/main/java/dev/yanshouwang/core/widget/BubbleMenu.java
@@ -1,0 +1,4 @@
+package dev.yanshouwang.core.widget;
+
+public class BubbleMenu {
+}
