+++ core/src/main/res/drawable/menu_arrow_vector.xml
