+++ core/src/main/res/layout/menu_layout.xml
