+++ core/src/main/res/layout/navigation_item_merge.xml
