+++ core/src/main/res/values/styles.xml
