+++ demo/src/main/java/dev/yanshouwang/core/demo/MainActivity.java
@@ -1,0 +1,14 @@
+package dev.yanshouwang.core.demo;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
