--- demo/src/main/java/dev/yanshouwang/wonder/demo/MainActivity.java
@@ -1,14 +1,0 @@
-package dev.yanshouwang.wonder.demo;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
