+++ settings.gradle
@@ -1,2 +1,2 @@
-include ':demo', ':wonder', ':core'
+include ':demo', ':core'
 rootProject.name='Wonder'
