--- wonder/src/main/java/dev/yanshouwang/wonder/widget/BubbleMenu.java
@@ -1,4 +1,0 @@
-package dev.yanshouwang.wonder.widget;
-
-public class BubbleMenu {
-}
